feat: add builder for cache client
To use in multithreaded code, we need a clonable type for
constructing new cache clients.

This diff adds a new `SimpleCacheClientBuilder` which contains the
channels that will be used to communicate with the gRPC endpoints.
Since these channels can be cloned at a low cost, we can clone the
builder and use it to build a new client from the cloned channels.