feature: allow changing the builder default ttl
Allow the `SimpleCacheClientBuilder` default ttl to be modified by
adding a new function. This is necessary if we want to be able to
create clients with different ttls after initializing the builder.

For example, this is desirable in the proxy where we want to allow
a different ttl for each cache to be used in place of no ttl.