feat: add auth login() binding
This adds an auth module. The flagship functionality is the login()
method. This method is the bridge between the grpc login message
stream and login implementations; to wit, `momento` looms large in
one's mind.

These types are a bit obtuse, and Rust is a little nascent in its
support of a couple things that would allow me to make a cleaner
api. Asynchronous generators and yield syntax being moved into
the `stable` release would do nice things for this shape. As a
consequence of this, testing beyond doctest modeling has proven
to be more of an effort than it is worth - particularly as the
service implementation does not yet exist.

As it is, the doctest illustrates the 2 facets of a login:
* LoginAction - actions to be taken by an integration during a login
* LoginResult - the final outcome of the login.
The actor method that an integration supplies to a login simply follows
the instructions sent down by the Momento login server, and the caller
of login() simply uses the result of the login, whether successful or
not.

Note that there are no unwrap()'s or expect()'s. The only panics that
shall rise out of a login() invocation will be bugs or fatal
environment errors beyond our control.