chore: Align MomentoCacheListResponse and list_caches (#56)
* chore: Align MomentoCacheListResponse and list_caches

The method `SimpleCacheClient::list_caches` supports pagination. It
does this by means of the parameter `next_token`: the client starts
iteration with this set to `None`, and ends when
`MomentoCacheListResponse::next_token` is `None`.

Previously `MomentoCacheListResponse::next_token` was not optional, so
the end sentinel was empty string as opposed to `None`. This means the
client would have to start iteration with `None` and test for end
iteration with empty string, which is confusing.

* Fix list_caches to run in a loop.

Fixes problems from the previous commit re: the type and ownership of
`next_token` in `list_caches`. Updated unit test to test looping
behavior.

* Fix formatting.

* Appease clippy