feat: Flatten `momento::response` module into a flat public API (#90)
The response module has a whole bunch of submodules that don't really add much information and make things much harder to find and use. This PR flattens the public API down into just the response module while still keeping the other modules for code organization internally.

* Rename examples -> example

If the folder is named examples then cargo thinks that any rust sources
in there are a standalone example and ignores the Cargo.toml in there.