feat: Move `SimpleCacheClient` and associated types to crate root (#96)
* feat: Export contents of simple_cache_client module in the root
* Export MomentoError in the crate root

make the `simple_cache_client` module private and export everything it used to export in the root module. `SimpleCacheClient` is the main type that users of the crate will be interacting with so it should really be in the root module. I have also re-exported `MomentoError` in the root module and added a `MomentoResult` type as a shorthand for `Result<T, MomentoError>`.