feat: Make TTL parameters use more idiomatic Rust
This commit does two main things:
- Use Duration for all time-span values with the exception of
  ttl_minutes when creating a signing key.
- Introduce a TtlPolicy type and use that for all TTL parameters in
  cache client methods where a TTL was previously taken. This is similar
  to the CollectionTtl type in the JS SDK but I've used it a bit more
  widely.

This means that all time-spans now go through std's Duration type and
there is now now way that users could, for example, pass seconds to a
method that is expecting milliseconds. The one exception to this for
create_signing_key. Duration has no from_minutes method and I felt that
having the duration silently truncate down to the last minute would be
too unintuitive if the only constructor is Duration::from_secs.