feat: add feature preview for topics
Momento topics are currently under development but this first step
helps enable end-user experimentation with the types.

This adds a Pubsub client, complete with connect(), publish() and
subscribe().

subscribe() is the more interesting method, as it returns a
Subscription, which is a new kind of user-facing functionality for
our sdk. I've wrapped the protobuf types into simpler non-Optional
messages, marshaling some of their safe/default notions into what
they mean for a Momento subscription.