chore: revise error types (#119)
Adds special handling for the very very special no-error kinds of
http2 errors. This change digs into the root cause of some special
tonic statuses and exposes simple momento error kinds. Specifically,
we're adding `Interrupted` as a probably-not-quite-an-error error to
wrap the HTTP2 notion of terminating requests without error. It might
be appropriate to treat an `Interrupted` as an error in some cases,
while generally it's not an error from a networked systems perspective.

Also this change attaches `source`s to MomentoErrors, so users can
actually log Debug info that may be helpful. This is different from
the Display implementation, which is still high level / not very
informative :-).