feat: pin to working protos version, add lockfile
A patch release of the protos contained a breaking change, and
because we didn't pin to a specific version or have a Cargo.lock
file, this caused the build to break for this project and
consuming projects.  This commit adds the Cargo.lock file,
and (for now) pins to a version of the client protos that this
library will compile successfully against.