chore: get rid of chrono dependency
Prior to this commit we had a dep on a time library called chrono.

In this PR we removed the Cargo.lock file from .gitignore, because
without it we were picking up patch releases of dependencies that
could cause the build of HEAD to break without any changes to our
code.

When Cargo.lock was re-added, github found a security vulnerability
in the `time` library that `chrono` depended on.  There is no new
version of chrono that isn't pinned to a vulnerable version of `time`.

We also don't really need to force a dependency on chrono on our users.

To all of those ends, this commit removes the chrono dep in favor
of the stdlib SystemTime.