feat: pin to working protos version, add lockfile (#124)
* feat: pin to working protos version, remove chrono dependency

A patch release of the protos contained a breaking change.
This pins to a version of the client protos that this
library will compile successfully against.

Also in this PR, github found a  security vulnerability
in the `time` library that `chrono` depended on.  There is no new
version of chrono that isn't pinned to a vulnerable version of `time`.

We also don't really need to force a dependency on chrono on our users.

To all of those ends, this commit removes the chrono dep in favor
of the stdlib SystemTime.