feat: structure response::Get (#142)
* feat: structure response::Get

There are a couple updates in here.
First up, I'm removing the momento::response::MomentoGetStatus
enum entirely.

Next, I'm turning momento::response::MomentoGetResponse into an
enum. In the migration I'm removing "ERROR" from the enumerated
response states. We have MomentoResult, which encapsulates Error
already. Doubling the places to check for errors is not good!

In addition to the removal, I'm changing the naming convention
from scream case to the more normal Rust enum title case.

While looking at the names, I observed that there is no difference
in the ambiguity of these 2 naming options:
* momento::response::MomentoGetResponse
* momento::response::Get

however, the latter is much lighter syntactically and offers
familiar options for importing at whatever level of disambiguation
one might desire. It also does not repeat itself.

Rather than migrating the as_string() function of MomentoGetResponse,
which was actually more of an as_str(), I've implemented a couple of
TryFrom specializations. These are for users to easily cast responses
as hits in String and Vec<u8> and let errors propagate.
This allows people to choose if they want to treat Miss as Error.

In deciding not to migrate as_string(), I'm also closing off an
implicit behavior we had, where a non-utf-8 string would have been
silently treated as a Hit( "" ). Rather than clobber cache items,
the new response::Get's TryInto<String> and &str implementations
return a new momento::MomentoError::TypeError with the associated
cause wired up in the Rust way for nice tracebacks.

For users that match on hit/miss, I have added a GetValue for which
String has a TryFrom implementation and Vec<u8> is covered via From.
Documentation on the momento::response::Get demonstrates options for
parsing our responses, depending on the user's needs. I've included
doctests to help users know what their easy options are.

Also do the same thing for DictionaryGet
Also do the same thing for DictionaryFetch

The value type between DictionaryFetch and DictionaryGet is the same,
so I moved it out into a DictionaryPairs struct. In this way the value
types will be more familiar and easier to work with across dictionary
apis.