feat: prototype of new API using requests as builders
This is a prototype of the new API that Kenny, Nate and I have
been discussing. In this model we define a Request and a Response
struct for each API. The Request structs can be used as builders,
to support optional arguments and future additions to the API.
The CacheClient exposes the simplest possible syntax for each API,
with no optional arguments provide, and implements that by
instantiating one of the Request structs.

The Request trait requires each Request type to implement a `send`
function, to do the work of sending the request. This will allow
us to keep the logic for the implementation details of each API
in its own struct, rather than having all of the implementations
live direction in `CacheClient`.

`CacheClient` also provides a `send_request` fn that can be called
directly, accepting a `Request` type as an argument. This is the
lower-level API that callers will use if they need to specify
any of the optional arguments for any given request.