chore: move prep_request out of SimpleCacheClient (#174)
As we experiment with different ways to express the requests
and optional arguments, some of the request preparation may
move outside of the SimpleCacheClient class. The prep_request
fn used by all of the request API fns was a member fn of
SimpleCacheClient, even though it didn't reference `self`. This
commit moves it out of the struct and into a regular function
so that it will be accessible from outside SimpleCacheClient.
This will allow us to more freely experiment with creating
requests from other places in the code.