feat: prototype of new API using requests as builders (#175)
* chore: move prep_request out of SimpleCacheClient

As we experiment with different ways to express the requests
and optional arguments, some of the request preparation may
move outside of the SimpleCacheClient class. The prep_request
fn used by all of the request API fns was a member fn of
SimpleCacheClient, even though it didn't reference `self`. This
commit moves it out of the struct and into a regular function
so that it will be accessible from outside SimpleCacheClient.
This will allow us to more freely experiment with creating
requests from other places in the code.

* feat: prototype of new API using requests as builders

This is a prototype of the new API that Kenny, Nate and I have
been discussing. In this model we define a Request and a Response
struct for each API. The Request structs can be used as builders,
to support optional arguments and future additions to the API.
The CacheClient exposes the simplest possible syntax for each API,
with no optional arguments provide, and implements that by
instantiating one of the Request structs.

The Request trait requires each Request type to implement a `send`
function, to do the work of sending the request. This will allow
us to keep the logic for the implementation details of each API
in its own struct, rather than having all of the implementations
live direction in `CacheClient`.

`CacheClient` also provides a `send_request` fn that can be called
directly, accepting a `Request` type as an argument. This is the
lower-level API that callers will use if they need to specify
any of the optional arguments for any given request.

* fix: Address PR feedback to remove one level of generics from MomentoRequest

* fix: linting

* fix: clippy

* fix: linting

* fix: env vars for doc tests

* chore: remove redundant trait