chore: update CONTRIBUTING re: tests
The goal of this commit is to update the CONTRIBUTING.md file to
include more information about how the different kinds of tests
should be organized and how to run them.

I also tweaked the directory structure of the int tests again. I'm
flattening it back out because unfortunately the `cargo test` CLI
doesn't seem to have a good way to only run selected tests if they
are nested in subdirs, though it has good support for it when they
are defined in the root of the `tests` dir.