feat: add `IntoSortedSetElements` trait and implementations for Vec and HashMap (#207)
* feat: adds a `IntoSortedSetElements` trait

This adds a `IntoSortedSetElements` trait to serve as an interface
between various collection types and the
`SortedSetPutElementsRequest`.

We provide implementations for `Vec` and `HashMap` and refactor the
code to consume the trait.

* chore: update docstrings

* fix: clobbered parameter type

* fix: fix doctests

* chore: change visibility `map_and_..._set_elements` pub -> private

* fix: do not run doctests on private helper

Since the doctests run externally, they need public visiblity to run.

* fix: exclude utility function from doctests

* chore: use `IntoSortedSetElements` in cache api

* refactor: introduce `SortedSetElement` external-facing type

Refactors the `IntoSortedSetElements` trait to use an
externally-facing `SortedSetElement` type. This allows users to create
a `Vec<SortedSetElement>` and use that in the request.

* fix: decouple sorted_set_name and sorted set element value types

Previously we had bound the trait for `sorted_set_name` and
`IntoSortedSetElements` generic type to the same thing. This is
undesirable since the `sorted_set_name` and value type for the latter
don't _have_ to be the same type. They both need to be `IntoBytes`.

* chore: use std::marker::PhantomData

* doc: document phantomdata usage

* test: add various integration tests for sorted set put elements

* refactor: dry test

* fix: broken integration test

* chore: name twiddle

* chore: remove unnecessary clones on test cache_name

* chore: use `as_str` to pass sorted_set_name in tests

---------

Co-authored-by: anitarua <anita@momentohq.com>