test: add test data types to ease test data creation (#233)
A common concern in the integration tests is creating test data. In
Rust, creating test data is a bit more painful than in other
languages. Thus we introduce some helper structs to create test data.

These test data create random data for each data type we have. In the
future we can also add comparators (PartialEq implementations) which
will be a common concern and pain in the test code.