test: more scalar coverage and test conveniences (#235)
We audit test coverage for get, set, and sorted sets and add tests as needed. We also implement traits and conveniences that will make future integration test authorship easier. In particular:

For each test data type, we implement the From trait paired with the corresponding cache hit response. That way we can use the test data directly in an assertion without the need to massage it in the right form.
We extract asserting a collection fetch is equal to test data. Both unpacking the actual result, massaging the expected reuslt, and comparing contain a lot of boilerplate. The sorted set tests demonstrate this pattern.