refactor: do not use trait
Unfortunately while a trait can capture the idea of an interface,
callers need to explicitly use the trait for this to work. Previously
the callers would have to `use momento::cache::configurations::{self,
PrebuildConfiguration}` to do this, which seemed weird.

For this reason we back out of the trait and instead define the
functions on each of the structs.