feat: implement custom debug and display traits for byte-array types (#323)
Presents a pattern we can use to get debug and display for response types.

Because our cache stores data as bytes, when users build a request or receive a response, they get back a wrapper around Vec<u8> of some form. This may be a Vec<u8> directly or another data structure like a HashMap or another Vec, for example.

In order to aid debugging and printing, we provide debug and display trait implementations that attempt to interpret and display the bytes as strings. There are two aspects to this:

For each Value type in a response, we manually implement std::fmt::Debug and std::fmt::Display. We have authored utility functions in utils::fmt to aid with this. See the get_response::Value implementation.
We auto-derive Debug and Display for the response type (GetResponse in this PR). Unfortunately Rust does not allow one to auto-derive Display, so we had to use the third-party library derive_more for that. Hence this PR adds that as a dependency.
Going forward, we can implement debug and display as appropriate for the Value types and auto-derive Debug and Display for the response types. We can also add truncation in the utility functions.

Closes #282