chore: implement Debug for DictionaryFetchResponse, tweak debug fmt utils
I wanted to play with what it would feel like to implement Debug for
other response types. In this PR I implemented it for DictionaryFetchResponse.

However, along the way I found that it was really hard to get the pretty-print
indenting correct when we are manually calling `Formatter::write*`; it's much
easier if we can stick to calling the `Formatter::debug_*` variants. So
I refactored the `fmt` utils a bit, to give us a `DebuggableValue` trait. This
trait provides an `as_debuggable_value` function for vec<u8>, String, and &str.
The implementations of Debug for the response Value structs can just use this
to get an instance of DebuggableValue, and then use `Formatter.debug_*` to
debug whatever collection type is appropriate.