fix: publish task broken with sdk subdir
In a recent commit we moved the sdk code into a subdir to avoid
Cargo workspace collisions. That caused the publish to fail
because it expects a README to exist in the directory we're publishing
from.

This commit moves the publish logic into the Makefile to centralize it
like the other build tasks, and adds a step to copy the README from
the parent directory before we execute the publish.