chore: add sdk/README to gitignore to prevent attempt to publish from dirty tree
In an earlier commit we moved the sdk to a subdir to prevent issues
with Cargo workspaces related to having a Cargo.toml file in a
parent dir of the examples.

We are now trying to publish the SDK from this new `sdk` subdir,
but it needs a README.md, so we are copying that file from the root
dir of the github repo into the `sdk` dir prior to publishing.

However, the `cargo publish` operation doesn't allow publishing
from a dirty git working tree. This commit adds the README to
the sdk/.gitignore so that the tree won't be flagged as dirty.

Another option we could consider in the future would be to add
another README generator action to the github actions, to just
generate a second copy of the root README and output it to the
`sdk` directory.