Align PR with the latest API design
Update to latest protos.

Switch get to success instead of hit/miss. This was in error before
because it assumed that a miss would result in an empty returned object
instead of an error.

Rename set to put.

Add a line to the client doc stating that the API is currently subject
to change.