Update tests for new Get response and Not Found Errors
Replace instances of NotFoundError in the tests and doctests with the
appropriate new not found error types.

Replace GetResponse::Success with a bare GetResponse in tests and
doctests.

Replace the NotFoundErrors that were returned by a Get miss with a
GetResponse containing None.

Remove the temporary example and revert the local sdk version used in
the examples directory.

Fix lint issues.