fix: interceptor should send only once per client (#384)
Previously we used a `static` variable to keep track of whether the
agent header was sent. This would send one agent header
unconditionally of the client type.

In order to send a single agent header per client type (cache vs
topics vs store) we make the variable an instance variable of the
`HeaderInterceptor`.