chore: add direct connect handshake
Momento direct connect uses an http+json -> protosocket negotiation
handshake. The http connection retrieves a list of socket addresses
that can be used with the connection-oriented protosocket rpc
protocol.

The ProtosocketCacheClient queries the address list on initial
connection, and updates the local list in the background. When a new
connection is required, for example due to a disconnect, the local
list is used to choose a new socket address for that "slot."

Direct connections are lighter than grpc connections, but currently
have some compromises. For example, they do not currently support
graceful GOAWAY mechanism. Also, the cached local address list may
be out of date with reality for a few seconds, causing some possible
re-connect attempts to fail until list refresh.

This connection scheme will be improved over time, and is expected
to outperform grpc for both cost & latency now, and match grpc
practically for resilience in the future.